requirements.txt
flask
flask-cors
insightface
onnxruntime
numpy
pillow
pandas
openpyxl
scikit-learn